fixed bug with loads from not root. migrated feature class to song_classes.py
